table type change

+++ database/migrations/2022_03_01_093561_create_payrolls_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('timesheet_id');
             $table->date('day_weekend');
-            $table->integer('total_amount');
+            $table->float('total_amount');
             $table->unsignedBigInteger('status_id')->default(1) ;
             $table->timestamps();
